+++ lib/src/controls/surfaces/flyout/flyout.dart
@@ -158,7 +158,7 @@ class Flyout extends StatefulWidget {
   final NavigatorKey? navigatorKey;
 
   @override
-  _FlyoutState createState() => _FlyoutState();
+  State<Flyout> createState() => _FlyoutState();
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {